JG :: seprating cart page UI decorate functions
