+++ data/pages/compositions/cart.html
@@ -317,4 +317,4 @@
     </div>
 </div>
 
-<script defer src="../web-resources/js/productUtilities.js" type="module"></script>
+<script defer src="../web-resources/js/decorators/decorate-cart.js" type="module"></script>
